changed float to decimal

+++ TddCalculator/RpnCalculator/RpnBinaryOperations.fs
@@ -1,10 +1,10 @@
 ﻿module RpnBinaryOperations
 
-let add x y : float = x + y
-let subtract x y : float = x - y
-let multiply x y : float = x * y
+let add x y : decimal = x + y
+let subtract x y : decimal = x - y
+let multiply x y : decimal = x * y
 
-let divide x y : float =
+let divide x y : decimal =
     match y with
-    | 0. -> raise (invalidOp "Cannot divide by zero")
+    | 0m -> raise (invalidOp "Cannot divide by zero")
     | _ -> x/y
